+++ src/plugin/core/net/net.h
@@ -32,7 +32,7 @@ class Net : public INet {
     }
 
     template <typename BaseType>
-    Net(BaseType *pBaseType, void (BaseType::*handleReceive)(const socket_t, const int, const char *, const uint32_t),
+    Net(BaseType *pBaseType, void (BaseType::*handleReceive)(const socket_t, const uint32_t, const char *, const uint32_t),
         void (BaseType::*handleEvent)(const socket_t, const SQUICK_NET_EVENT, INet *), bool tcpStream = false) {
         mxBase = NULL;
         listener = NULL;
@@ -62,8 +62,8 @@ class Net : public INet {
     virtual bool SendData(const char *data, const size_t len, const socket_t sock) override;
     bool SendDataToAllClient(const char *data, const size_t len) override;
 
-    virtual bool SendMsg(const int16_t msg_id, const char *msg, const size_t len, const socket_t sock) override;
-    virtual bool SendMsgToAllClient(const int16_t msg_id, const char *msg, const size_t len) override;
+    virtual bool SendMsg(const int32_t msg_id, const char *msg, const size_t len, const socket_t sock) override;
+    virtual bool SendMsgToAllClient(const int32_t msg_id, const char *msg, const size_t len) override;
 
     virtual bool CloseNetObject(const socket_t sock) override;
     virtual bool AddNetObject(const socket_t sock, NetObject *pObject) override;
@@ -75,7 +75,7 @@ class Net : public INet {
     virtual int GetConnections() override;
 
   private:
-    bool SendMsg(const int16_t msg_id, const char *msg, const size_t len, const std::list<socket_t> &fdList);
+    bool SendMsg(const int32_t msg_id, const char *msg, const size_t len, const std::list<socket_t> &fdList);
     bool SendData(const char *msg, const size_t len, const std::list<socket_t> &fdList);
 
   private:
@@ -100,7 +100,7 @@ class Net : public INet {
 
   protected:
     int DeCode(const char *strData, const uint32_t ulen, rpcHead &xHead);
-    int EnCode(const uint16_t umsg_id, const char *strData, const uint32_t unDataLen, std::string &strOutData);
+    int EnCode(const uint32_t umsg_id, const char *strData, const uint32_t unDataLen, std::string &strOutData);
 
   private:
     //<fd,object>
